mali_kbase: pm: Enhance power policy interface
Extends the power management policy interface to allow
for more complex power policies to be implemented by:

(1) Adding a 'handle_event' function which can accept
incoming power events from the GPU power state machine.
Also updates the GPU power state machine implemented in
 to emit such events.

(2) Allow for the power policy to modify the number of
ticks that the shader tick timer runs for. This is
implemented by adding a new field element that retains
the compile time & sysfs updated value for the number
of ticks, leaving the existing
element to be modified by the power policy. For simple
power policies, this results in no change.

(3) Add logic to automatically clear per-power policy
private data when switching between power policies.
This reduces the amount of cleanup that each power
policy needs to do on init/term.

These allow for more advanced power policies to be
implemented.

Bug: 183700775
Co-authored-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: I81766b7c97691fec2496f60d91b00476277b2fe7
Signed-off-by: Tkpointz <pointzztk@gmail.com>
[cyberknight777: Port to mali_bifrost]
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: SakthivelNadar <s2234nadar@gmail.com>