BACKPORT: HACK: bpf: port bpf functions list from 4.19
Android 14 QPR1 will require kernel 4.19. However, the current beta only
requires a couple of function more compared to Android 14.

These functions are number 68 and 125, so we don't actually need all of
the functions in between. To simplify the backporting, we can just copy
here the functions from the 4.19 Android stable branch and, when needed,
actually define the ones we need.

This does not cause issues as without definition the bpf verifier will
return an error without crashing.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>